requirements.txt
beautifulsoup4==4.12.0
pandas==2.0.3
sqlalchemy==2.0.0
requests==2.28.2
python-dotenv==1.0.0
openai==1.0.0
selenium==4.10.0
schedule==1.2.0
numpy==1.24.3